Cleanup utils package in x/auth/client
