+++ x/auth/client/utils/cmd_test.go
@@ -1,0 +1,50 @@
+package utils_test
+
+import (
+	"testing"
+
+	"github.com/spf13/cobra"
+	"github.com/stretchr/testify/require"
+
+	"github.com/cosmos/cosmos-sdk/x/auth/client/utils"
+)
+
+func TestValidateCmd(t *testing.T) {
+	// setup root and subcommands
+	rootCmd := &cobra.Command{
+		Use: "root",
+	}
+	queryCmd := &cobra.Command{
+		Use: "query",
+	}
+	rootCmd.AddCommand(queryCmd)
+
+	// command being tested
+	distCmd := &cobra.Command{
+		Use:                        "distr",
+		DisableFlagParsing:         true,
+		SuggestionsMinimumDistance: 2,
+	}
+	queryCmd.AddCommand(distCmd)
+
+	commissionCmd := &cobra.Command{
+		Use: "commission",
+	}
+	distCmd.AddCommand(commissionCmd)
+
+	tests := []struct {
+		reason  string
+		args    []string
+		wantErr bool
+	}{
+		{"misspelled command", []string{"comission"}, true},
+		{"no command provided", []string{}, false},
+		{"help flag", []string{"comission", "--help"}, false},
+		{"shorthand help flag", []string{"comission", "-h"}, false},
+	}
+
+	for _, tt := range tests {
+		err := utils.ValidateCmd(distCmd, tt.args)
+		require.Equal(t, tt.wantErr, err != nil, tt.reason)
+	}
+}
