+++ x/auth/client/utils/errors.go
@@ -2,4 +2,7 @@ package utils
 
 import "errors"
 
-var errInvalidSigner = errors.New("tx intended signer does not match the given signer")
+var (
+	errInvalidSigner        = errors.New("tx intended signer does not match the given signer")
+	errInvalidGasAdjustment = errors.New("invalid gas adjustment")
+)
